--- src/components/sidebar.css
@@ -1,49 +1,0 @@
-.sidebar {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 20%;
-  height: 100vh;
-  border-right: 1px solid #fafafa;
-  box-shadow: 0 0 3px 0;
-  gap: 0;
-  padding: 0 1.9% 0 0;
-}
-
-.nav-logo {
-  width: 45%;
-  margin-top: 20px;
-}
-
-.nav-box {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  flex-direction: row;
-  width: 100%;
-  height: 50px;
-  font-size: 13px;
-  text-decoration: none;
-  color: #000;
-  padding: 0 5%;
-}
-
-.nav-box:hover {
-  color: #fff;
-  background-color: #97bf0f;
-  font-size: 11px;
-}
-
-.social-box {
-  position: fixed;
-  bottom: 20px;
-  left: 4%;
-}
-
-.social-link {
-  color: #000;
-}
-
-.social-link:hover {
-  color: #97bf0f;
-}
